Fix messages of command: rm
Right now, rm command doesn't differentiate between a file and a directory.
For example, on deleting a directory, it says "file deleted". This patch
fixes the issue.

+++ crash/src/main/resources/crash/commands/cloudius/rm.groovy
@@ -22,10 +22,18 @@ class rm implements Completer {
     if (!file.exists()) {
       throw new ScriptException("no such file or directory")
     }
-    if (!file.delete()) {
-      throw new ScriptException("cannot delete file")
+    if (file.isDirectory()) {
+      if (!file.delete()) {
+        throw new ScriptException("cannot delete directory")
+      }
+      out.println("directory deleted")
     }
-    out.println("file deleted")
+    if (file.isFile()) {
+      if (!file.delete()) {
+        throw new ScriptException("cannot delete file")
+      }
+      out.println("file deleted")
+    }
   }
 
   @Override
